@@ -1,6 +1,6 @@
 {
   "name": "react-redux-firebase",
-  "version": "2.4.1",
+  "version": "2.5.0",
   "description": "Redux integration for Firebase. Comes with a Higher Order Components for use with React.",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -63,7 +63,7 @@
     "prop-types": "^15.7.2"
   },
   "peerDependencies": {
-    "react": "^0.14.6 || ^15.0.0-0 || ^16.0.0-0"
+    "react": "^16.3.0"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
@@ -84,14 +84,14 @@
     "babel-register": "^6.24.0",
     "chai": "^3.5.0",
     "chai-as-promised": "^7.1.1",
-    "chai-enzyme": "^0.8.0",
+    "chai-enzyme": "^1.0.0-beta.1",
     "child-process-promise": "^2.2.1",
-    "codecov": "^3.5.0",
+    "codecov": "^3.6.1",
     "cross-env": "^5.1.0",
-    "docdown": "^0.7.2",
     "documentation": "^5.3.2",
     "documentation-markdown-api-theme": "^1.0.2",
-    "enzyme": "^2.3.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.14.0",
     "eslint": "^4.8.0",
     "eslint-config-prettier": "^2.3.0",
     "eslint-config-standard": "^10.2.1",
@@ -112,9 +112,8 @@
     "lodash-webpack-plugin": "^0.11.5",
     "mocha": "^3.5.3",
     "prettier": "1.10.2",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
     "recompose": "^0.26.0",
     "redux": "4.0.0",
     "redux-firestore": "0.5.7",
